refactor: 폴더 구조 변경 - 각 도메인 별로 분리하고 가까운 파일들 끼리 폴더화
