+++ src/advanced/__tests__/advanced.test.tsx
@@ -2,8 +2,8 @@ import { act, fireEvent, render, screen, within } from '@testing-library/react';
 import { useState } from 'react';
 import { describe, expect, test } from 'vitest';
 
-import { AdminPage } from '../../refactoring/components/domains/admin/AdminPage';
-import { CartPage } from '../../refactoring/components/domains/cart/CartPage';
+import { AdminPage } from '../../refactoring/domains/admin/AdminPage';
+import { CartPage } from '../../refactoring/domains/cart/CartPage';
 
 import type { Coupon, Product } from '../../types';
