+++ src/refactoring/domains/admin/components/coupon/CouponManager.tsx
