+++ src/refactoring/domains/admin/components/product/ProductItem.tsx
