+++ src/refactoring/domains/admin/components/product/ProductItemDetail.tsx
