+++ src/refactoring/domains/admin/components/product/ProductManager.tsx
