+++ src/refactoring/domains/cart/hooks/useCart.ts
@@ -1,9 +1,8 @@
-// useCart.ts
 import { useState } from 'react';
 
-import { calculateCartTotal, updateCartItemQuantity } from '../models/cart';
+import { calculateCartTotal, updateCartItemQuantity } from '../../../models/cart';
 
-import type { CartItem, Coupon, Product } from '../../types';
+import type { CartItem, Coupon, Product } from '../../../../types';
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
